False values are real values
With the previous implementation if you tried to put a value of `false`
in configuration (such as for the `:registry_schemas_autoreg` key) it
would always be `true`. The only way to get a `false` value is to set it
in the `use Avrora.Client` expression.

This change allows us to set `false` (or any other falsy value) in our
configuration.